HBAPP-95 Move, rename files
