+++ .env.example
@@ -12,7 +12,7 @@ AUTH_GITHUB_SECRET=XXXXXXXX
 # Set the following only when deployed. In this example, we can reuse the same OAuth app, but if you are storing users, we recommend using a different OAuth app for development/production so that you don't mix your test and production user base.
 # AUTH_REDIRECT_PROXY_URL=https://YOURAPP.vercel.app/api/auth
 
-# Instructions to create kv database here: https://vercel.com/docs/storage/vercel-kv/quickstart and 
+# Instructions to create kv database here: https://vercel.com/docs/storage/vercel-kv/quickstart
 KV_URL=XXXXXXXX
 KV_REST_API_URL=XXXXXXXX
 KV_REST_API_TOKEN=XXXXXXXX
