+++ app/page.tsx
@@ -1,7 +1,7 @@
-import { AppContainer } from '@/components/app-container'
+import { AppEntry } from '@/components/features/pages/page-app-entry'
 
 export const runtime = 'edge'
 
 export default function IndexPage() {
-  return <AppContainer />
+  return <AppEntry />
 }
