+++ components/features/chat/chat-list.tsx
@@ -1,7 +1,7 @@
 import { type Message } from 'ai'
 
 import { Separator } from '@/components/ui/separator'
-import { ChatMessage } from '@/components/chat-message'
+import { ChatMessage } from '@/components/features/chat/chat-message'
 
 export interface ChatList {
   messages: Message[]
