+++ components/features/chat/chat-message.tsx
@@ -6,7 +6,7 @@ import { cn } from '@/lib/utils'
 import { CodeBlock } from '@/components/ui/codeblock'
 import { MemoizedReactMarkdown } from '@/components/markdown'
 import { IconOpenAI, IconUser } from '@/components/ui/icons'
-import { ChatMessageActions } from '@/components/chat-message-actions'
+import { ChatMessageActions } from '@/components/features/chat/chat-message-actions'
 
 export interface ChatMessageProps {
   message: Message
