+++ components/features/chat/interface-cloud-chat.tsx
@@ -4,12 +4,12 @@ import { useEffect } from 'react'
 import { useChat, type Message } from 'ai/react'
 import { useTheme } from 'next-themes'
 import { cn } from '@/lib/utils'
-import { ChatList } from '@/components/chat-list'
-import { ChatPanel } from '@/components/chat-panel'
-import { EmptyScreen } from '@/components/empty-screen'
-import { ChatScrollAnchor } from '@/components/chat-scroll-anchor'
+import { ChatList } from '@/components/features/chat/chat-list'
+import { ChatPage } from '@/components/features/chat/chat-page'
+import { EmptyScreen } from '@/components/features/chat/chat-empty-screen'
+import { ChatScrollAnchor } from '@/components/features/chat/chat-scroll-anchor'
 import { toast } from 'react-hot-toast'
-import { useSettings } from './features/settings/hooks'
+import { useSettings } from '../settings/hooks'
 import { AIModels } from '@/components/features/settings/hooks'
 
 interface IProps extends React.ComponentProps<'div'> {
@@ -53,7 +53,7 @@ export function CloudChat({ id, initialMessages, modelId, className }: IProps) {
           <EmptyScreen setInput={setInput} />
         )}
       </div>
-      <ChatPanel
+      <ChatPage
         id={id}
         isLoading={isLoading}
         stop={stop}
