+++ components/features/menus/app-entry/menu-application-modes.tsx
@@ -7,8 +7,8 @@ import { QuestionMarkIcon, PersonIcon, ClipboardIcon } from '@radix-ui/react-ico
 import { buttonVariants } from '@/components/ui/button'
 import { IconPlus } from '@/components/ui/icons'
 import { Tabs } from '@/components/ui/tabs'
-import { Playground } from '@/components/features/menus/app/tab-playground'
-import { BotCreationMenu } from '@/components/features/menus/app/tab-bots'
+import { Playground } from '@/components/features/menus/app-entry/tab-playground'
+import { BotCreationMenu } from '@/components/features/menus/app-entry/tab-bots'
 import { I_ModelConfigs, I_ServiceApis, I_Text_Settings, T_InstalledTextModel } from '@/lib/homebrew'
 import { toast } from 'react-hot-toast'
 import { cn } from '@/lib/utils'
